Add beta_from_energy functionality and update CMake configuration
- Implement beta_from_energy function in new source files.
- Update CMakeLists.txt to include new library and link dependencies.
- Modify pyproject.toml to reflect author and license changes, and add classifiers.
