+++ pyproject.toml
@@ -7,10 +7,20 @@ name = "pyamtrack"
 dynamic = ["version"]
 description = "Python bindings for libamtrack"
 readme = "README.md"
-authors = [{ name = "Your Name", email = "your.email@example.com" }]
+authors = [{ name = "Leszek Grzanka", email = "grzanka@agh.edu.pl" }]
 requires-python = ">=3.9"
-license = { file = "LICENSE" }
+license = "GPL-3.0-or-later"
 dependencies = ["numpy"]
+classifiers = [
+  "Development Status :: 3 - Alpha",
+
+  # Specify the Python versions you support here.
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+]
 
 [tool.scikit-build]
 wheel.expand-macos-universal-tags = true
