+++ src/beta_from_energy.h
@@ -1,0 +1,8 @@
+#ifndef BETA_FROM_ENERGY_H
+#define BETA_FROM_ENERGY_H
+
+#include <pybind11/pybind11.h>
+
+pybind11::object PYBIND11_EXPORT beta_from_energy(pybind11::object input);
+
+#endif // BETA_FROM_ENERGY_H
